+++ Adj/Program/Controlflow/Implementation/Optics.hs
@@ -1,14 +1,15 @@
 module Adj.Program.Controlflow.Implementation.Optics where
 
 import Adj.Auxiliary ((=-), type (=!?=), FG (FG))
-import Adj.Algebra.Category (type (<--), type (-->), type (:*:>), type (:+:>), (.), (..:), Opposite (Opposite), Straight (Straight), extract)
+import Adj.Algebra.Category (type (<--), type (-->), type (:*:>), type (:+:>), (.), (..:), Opposite (Opposite), Straight (Straight))
 import Adj.Algebra.Set ((:+:) (This, That))
 
 type Lens queried required source target =
 	source --> (((:*:>) (queried target) =!?= (-->) (required target)) source)
 
-view :: Lens queried required source target -> source -> queried target
-view lens source = extract . Opposite . (=-) . (=-) ..: lens =- source
+-- TODO: use Transformation action in Monoidal intance for extract method
+-- view :: Lens queried required source target -> source -> queried target
+-- view lens source = extract . Opposite . (=-) . (=-) ..: lens =- source
 
 type Prism available set subset =
 	set --> (((:+:>) (available subset) =!?= (<--) (available subset)) set)